Migrates to new version of Resource Pack
This new version is backed by a REST implementation. We are introducing
resources created for the specific use cases we are trying to solve, as
well as establishing a pattern that authors may follow to add new
resources.

The maintainers list will be updated in the near future with the correct
long term maintainers.

Signed-off-by: David McCown <dmccown@chef.io>

Co-authored-by: Paul Welch <pwelch@chef.io>
Co-authored-by: Trevor Bramble <tbramble@chef.io>
Co-authored-by: Joshua Padgett <jpadgett@chef.io>