Add Integration Tests for CIS-5.x Resources
Add InSpec integration tests for the resources that were
added for CIS-5.x benchmarks.

Add TerraForm Output that can be used with new integration tests.

This also fixes the Azure API endpoint uses for Log Alerts. The API
was returning an error about a wrong API version number. This corrects
it to the one provided by the error message.

Fixes #80

Signed-off-by: Paul Welch <pwelch@chef.io>