Changes integration test and adds error handling
This changes the integration to check that we are getting data back when
using the plural resource. It also verifies that the right number of
guest accounts exist for your environment.

Adds some error handling so if you have an erorr such as "unauthorized"
then your test will fail with an error indicating that you don't have
sufficent privileges.

Signed-off-by: David McCown <dmccown@chef.io>