Clean up FilterTable definition, use it for Guests
In this resource pack the convention is to chain FilterTable
definitions, and to define a method with memoization for access to
`@table`.

I removed the nested `Guest` class because it should just be a `where`
query on the FilterTable-enabled resource. I left the convenience method
but I think that may be unconventional versus using the `where` query in
your controls.

Signed-off-by: Trevor Bramble <tbramble@chef.io>