Exempts tags from being turned into a struct
Turning tags into a struct makes it difficult to compare the expected
sets of tags with the actual set of tags returned. There's also the
potential for dashes in tag names. In these cases the struct methods
are only accessable by calling `send` on the struct with the key names.

This change preserves tags as a hash.

fixes #138

Signed-off-by: David McCown <dmccown@chef.io>