Add more graceful handling of missing response data (#155)
This is more scar tissue around Azure APIs not adhering to a response contract such that the key for a given item is not returned when there is no attached item of that type. (If a VM has no attached drives, it does not return `{... "drives": [] ...}`, it just does not include that element at all.)

When testing against our struct data with dot-notation instead of hash index references (`data.foo` vs. `data[:foo]`) a `NoMethodError` exception would be raised if `foo` wasn't included in the API response.

This change handles that scenario by returning `NullResponse` when there is a reference to an unknown key. It's not ideal, but it'll help while we get more feedback from our current model and decide whether a larger change is appropriate.

Signed-off-by: Trevor Bramble <tbramble@chef.io>